Move nico_data and create niconico pkg
